Reenable protocol consolidation for tests.
