+++ apps/explorer/test/explorer/celo/events/event_map_test.exs
@@ -109,4 +109,17 @@ defmodule Explorer.Celo.ContractEvents.EventMapTest do
       assert result.value == 420_420_420
     end
   end
+
+  describe "General functionality" do
+    test "event map should have definitions for all events" do
+      event_map_topic_set = EventMap.map() |> Map.keys() |> MapSet.new()
+
+      contract_events_topic_set =
+        EventTransformer.__protocol__(:impls)
+        |> then(fn {:consolidated, modules} -> Enum.map(modules, & &1.topic()) end)
+        |> MapSet.new()
+
+      assert MapSet.subset?(contract_events_topic_set, event_map_topic_set)
+    end
+  end
 end
